tweaks

+++ shell/probe/raindrop.ps1
@@ -3,33 +3,7 @@
 function Log {
     Param([String]$line)
 
-    Write-EventLog -LogName "Application" -Source "Prelude Probe Service" -EventId 0 -Message $line -ErrorAction Ignore
-}
-
-function Run {
-    Param([String]$Dat = "")
-
-    $Vst = New-Item -Path "$Dir\$(New-Guid).exe" -Force
-    $Headers = @{
-        'token' = FromEnv "PRELUDE_TOKEN"
-        'dos' = $Dos
-        'dat' = $Dat
-        'version' = "1.0"
-    }
-    $Response = Invoke-WebRequest -URI $Api -UseBasicParsing -Headers $Headers -MaximumRedirection 1 -OutFile $Vst -PassThru
-    $Test = $Response.BaseResponse.ResponseUri.AbsolutePath.Split("/")[-1].Split("_")[0]
-
-    if ($Response.BaseResponse.ResponseUri -contains "upgrade") {
-        Log "[P] Upgrade required" && exit 1
-    } elseif (-not $Test) {
-        return
-    } elseif ($CA -ne $Response.BaseResponse.ResponseUri.Authority) {
-        return
-    }
-
-    Log "[P] Running $Test [$Vst]"
-    $Code = Execute $Vst   
-    Run -Dat "${Test}:$Code"
+    Write-EventLog -LogName "Application" -Source "Prelude Detect" -EventId 0 -Message $line -ErrorAction Ignore
 }
 
 function Execute { 
@@ -62,8 +36,29 @@ $Dos = "windows-$Env:PROCESSOR_ARCHITECTURE"
 
 while ($true) {
     try {
-        Run
-        Remove-Item $Dir -Force -Recurse
-    } catch { Log $_ }
-    Start-Sleep -Seconds $Sleep
+        $Vst = New-Item -Path "$Dir\$(New-Guid).exe" -Force
+        $Headers = @{
+            'token' = FromEnv "PRELUDE_TOKEN"
+            'dos' = $Dos
+            'dat' = $Dat
+            'version' = "1.0"
+        }
+        $Response = Invoke-WebRequest -URI $Api -UseBasicParsing -Headers $Headers -MaximumRedirection 1 -OutFile $Vst -PassThru
+        $Test = $Response.BaseResponse.ResponseUri.AbsolutePath.Split("/")[-1].Split("_")[0]
+    
+        if ($Test) {
+            if ($CA -eq $Response.BaseResponse.ResponseUri.Authority) {
+                Log "[P] Running $Test [$Vst]"
+                $Code = Execute $Vst   
+                $Dat = "${Test}:$Code"
+            }
+        } elseif ($Response.BaseResponse.ResponseUri -contains "upgrade") {
+            Log "[P] Upgrade required" && exit 1
+        } else {
+            Remove-Item $Dir -Force -Recurse
+            Start-Sleep -Seconds $Sleep
+        }
+    } catch { 
+        Log $_ 
+    }
 }
